+++ app/views/components/views/shared/insights_component.rb
@@ -19,7 +19,6 @@ module Views
           Insights::VisibleQuery
             .new(relation: @insightable.insights)
             .resolve(insightable: @insightable)
-            .includes(:entity)
             .load
       end