+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - query objects
+- cache entities data for skipping redundant queries
 
 ## [1.3.7] - 2023-09-06
 ### Added